add backend: models, dtos, services, configs and controllers

+++ backend/src/main/java/com/ecommerce/valdivian/controller/UserController.java
@@ -1,9 +1,16 @@
 package com.ecommerce.valdivian.controller;
 
-
-import com.ecommerce.valdivian.model.User;
+import com.ecommerce.valdivian.dto.UpdateUserDTO;
+import com.ecommerce.valdivian.model.*;
+import com.ecommerce.valdivian.repository.CartRepository;
 import com.ecommerce.valdivian.repository.UserRepository;
+import com.ecommerce.valdivian.service.UserService;
+import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.security.access.prepost.PreAuthorize;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -14,39 +21,67 @@ import java.util.Optional;
 @RequestMapping("/api/users")
 public class UserController {
     private final UserRepository userRepository;
+    private final CartRepository cartRepository;
+    private final UserService userService;
 
 
+    @PostMapping("/register")
+    public ResponseEntity<User> registerUser(@RequestBody RegistrationBody userDto) {
+        return userService.registerUser(userDto);
+    }
+
+    @PostMapping("/login")
+    public ResponseEntity<LoginResponse> loginUser(@Valid @RequestBody LoginBody loginBody) {
+        String jwt = userService.loginUser(loginBody);
+        if (jwt == null) {
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST).build();
+        } else {
+            LoginResponse response = new LoginResponse();
+            response.setJwt(jwt);
+            return ResponseEntity.ok(response);
+        }
+    }
+
+    @PreAuthorize("hasRole('admin')")
     @GetMapping
-    List<User> all(){
+    public List<User> getAllUsers() {
         return userRepository.findAll();
     }
 
-    @PostMapping
-    User user (@RequestBody User newUser){
-        return userRepository.save(newUser);
-    }
 
     @GetMapping("/{id}")
-    Optional<User> id(@PathVariable Long id){
-        return userRepository.findById(id);
+    public ResponseEntity<User> getUserById( @PathVariable Long id) {
+        Optional<User> optionalUser = userRepository.findById(id);
+        return optionalUser.map(ResponseEntity::ok).orElseGet(() -> ResponseEntity.notFound().build());
     }
 
-    @DeleteMapping("/{id}")
-    void deleteUser(@PathVariable Long id){
-        userRepository.deleteById(id);
-    }
-
+    @PreAuthorize("hasRole('admin')")
     @PutMapping("/{id}")
-    User updateUser(@RequestBody User updatedUser, @PathVariable Long id) {
-        return userRepository.findById(id)
-                .map(user -> {
-                    user.setUsername(updatedUser.getUsername());
-                    user.setPassword(updatedUser.getPassword());
-                    user.setAddress(updatedUser.getAddress());
-                    user.setEmail(updatedUser.getEmail());
-                    return userRepository.save(user);
-                })
-                .orElseThrow(() -> new RuntimeException("User not found with id: " + id));
+    public ResponseEntity<User> updateUser(@PathVariable Long id,
+                                           @Valid @RequestBody UpdateUserDTO updatedUserDto) {
+        return userService.updateUser(id, updatedUserDto);
+    }
+
+    @PreAuthorize("hasRole('admin')")
+    @DeleteMapping("/{id}")
+    public ResponseEntity<Void> deleteUser( @PathVariable Long id) {
+        Optional<User> optionalUser = userRepository.findById(id);
+        if (optionalUser.isEmpty()) {
+            return ResponseEntity.notFound().build();
+        }
+        User user = optionalUser.get();
+        List<Cart> carts = cartRepository.findAllByUser(user);
+        carts.forEach(cart -> cart.setUser(null));
+        cartRepository.saveAll(carts);
+        cartRepository.deleteAllByUser(user);
+        userRepository.deleteById(id);
+
+        return ResponseEntity.noContent().build();
+    }
+
+    @GetMapping("/me")
+    public User getLoggedInUserProfile(@AuthenticationPrincipal User user) {
+        return user;
     }
 
 }
